worker/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "juno-worker"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "pgvector>=0.3.0",
    "openai>=1.30.0",
    "hdbscan>=0.8.38",
    "umap-learn>=0.5.6",
    "numpy>=1.26.0",
    "pydantic-settings>=2.0.0",
]
